requirements.txt
appdirs==1.4.3
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
oauthlib==2.0.2
packaging==16.8
py-trello==0.9.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
